requirements_render.txt
gradio
requests
numpy
pillow
